+++ nip/language_model_server/trainers/dpo.py
@@ -265,7 +265,7 @@ def train(config: LmTrainingConfig, dataset: Dataset, job_id: str, new_model_nam
         training_lora_config = LoraConfig(**config.training_lora_config.model_dump())
 
     # The maximum length for a W&B job name is 128 characters.
-    job_name = job_id[:128]
+    job_name = job_id[:127]
 
     dpo_config = DPOConfig(
         **config.dpo_config.model_dump(),